refactor: extract PlanView from erg-dashboard.jsx
Part 1/3 of issue #78. Moves the inline 'plan' tab (today-forward
prescriptions from status='planned' sessions) into a standalone
views/PlanView.jsx component, following the pattern of the other
extracted views.
